+++ collector/src/main/java/com/bitmovin/analytics/data/RandomizedUserIdIdProvider.kt
@@ -1,0 +1,9 @@
+package com.bitmovin.analytics.data
+
+import java.util.UUID
+
+class RandomizedUserIdIdProvider : UserIdProvider {
+    private val userId: String = UUID.randomUUID().toString()
+
+    override fun userId(): String = userId
+}