+++ collector/src/main/java/com/bitmovin/analytics/data/SecureSettingsAndroidIdUserIdProvider.kt
@@ -1,0 +1,8 @@
+package com.bitmovin.analytics.data
+
+import android.content.Context
+import android.provider.Settings
+
+class SecureSettingsAndroidIdUserIdProvider(val context: Context) : UserIdProvider {
+    override fun userId(): String = Settings.Secure.getString(context.contentResolver, Settings.Secure.ANDROID_ID)
+}