+++ collector/src/main/java/com/bitmovin/analytics/utils/Util.java
@@ -1,11 +1,9 @@
 package com.bitmovin.analytics.utils;
 
-import android.content.Context;
 import android.content.res.Resources;
 import android.media.MediaCodecInfo;
 import android.media.MediaCodecList;
 import android.os.SystemClock;
-import android.provider.Settings;
 import android.util.Pair;
 import java.net.URI;
 import java.util.ArrayList;
@@ -38,10 +36,6 @@ public class Util {
 
     public static String getUUID() {
         return UUID.randomUUID().toString();
-    }
-
-    public static String getUserId(Context context) {
-        return Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ANDROID_ID);
     }
 
     /**